Improved: Fix or Silence various warnings (OFBIZ-10701)
In order to detect potential issues early, the linting compiler option should be 
used by default and disabled with {{./gradlew -PXlint:none build}}.

Additionally it is important to reduce the number of warnings otherwise new 
warnings will remain unnoticed.

The warning related to the com.googlecode.concurrentlinkedhashmap dependency 
should resolves itself once we handle upgrade to Caffeine like proposed in 
OFBIZ-6747.

Thanks: Mathieu Lirzin

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-framework/trunk@1848673 13f79535-47bb-0310-9956-ffa450edef68

+++ framework/common/src/main/java/org/apache/ofbiz/common/qrcode/QRCodeServices.java
@@ -156,7 +156,9 @@ public class QRCodeServices {
                     typeMap.put("height", logoImageMaxHeight.toString());
                     Map<String, Map<String, String>> dimensionMap = new HashMap<>();
                     dimensionMap.put("QRCode", typeMap);
-                    Map<String, Object> logoImageResult = ImageTransform.scaleImage(logoBufferedImage, (double) logoBufferedImage.getWidth(), (double) logoBufferedImage.getHeight(), dimensionMap, "QRCode", locale);
+                    Map<String, Object> logoImageResult =
+                            ImageTransform.scaleImage(logoBufferedImage, logoBufferedImage.getWidth(),
+                                    logoBufferedImage.getHeight(), dimensionMap, "QRCode", locale);
                     logoBufferedImage = (BufferedImage) logoImageResult.get("bufferedImage");
                 }
                 BitMatrix newBitMatrix = bitMatrix.clone();
